+++ src/lib/jp2/t1/t1_ht/T1HT.cpp
@@ -88,10 +88,10 @@ void T1HT::preEncode(encodeBlockInfo *block, grk_tcd_tile *tile,
 			int32_t shift = 31 - (block->k_msbs + 1) - 11;
 			for (auto i = 0U; i < w; ++i) {
 				int32_t temp = block->tiledp[tileIndex];
-				int32_t t = (int32_t)((float)temp * block->inv_step_ht * (1<<shift));
+				int32_t t = (int32_t)((float)temp * block->inv_step_ht * (float)(1<<shift));
 				int32_t val = t >= 0 ? t : -t;
 				maximum = max((uint32_t)val, maximum);
-				int32_t sign = t >= 0 ? 0 : 0x80000000;
+				int32_t sign = t >= 0 ? 0 : (int32_t)0x80000000;
 				int32_t res = sign | val;
 				unencoded_data[cblk_index] = res;
 				tileIndex++;
@@ -218,7 +218,7 @@ void T1HT::postDecode(decodeBlockInfo *block) {
 			for (auto i = 0U; i < cblk_w; ++i) {
 				int32_t temp = *src;
 				int32_t val = (temp & 0x7FFFFFFF) >> shift;
-				tile_row_data[i] = (int32_t)((temp & 0x80000000) ? -val : val);
+				tile_row_data[i] = (int32_t)(((uint32_t)temp & 0x80000000) ? -val : val);
 				src++;
 			}
 			tile_data += dest_width;
@@ -229,7 +229,7 @@ void T1HT::postDecode(decodeBlockInfo *block) {
 			float *GRK_RESTRICT tile_row_data = (float*)tile_data;
 			for (auto i = 0U; i < cblk_w; ++i) {
 		       float val = (float)(*src & 0x7FFFFFFF) * block->stepsize;
-		       tile_row_data[i] = (*src & 0x80000000) ? -val : val;
+		       tile_row_data[i] = ((uint32_t)*src & 0x80000000) ? -val : val;
 			   src++;
 			}
 			tile_data += dest_width;
